requirements.txt
Flask==3.0.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
Werkzeug==3.1.3
gunicorn==21.2.0
python-dotenv==1.0.1
alembic==1.15.2 